backend/requirements.txt
asgiref==3.5.0
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
idna==2.10
iniconfig==1.1.1
MarkupSafe==2.0.1
packaging==20.9
pluggy==0.13.1
psycopg2==2.9.3
py==1.10.0
pycodestyle==2.6.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.4
Werkzeug==2.0.2
